Bugfix for GC root mishandling in memprof.c
When allocating the OCaml block for a Gc.Memprof info block, memprof.c uses
e->user_data as the root for the callstack. While e->user_data is indeed a root,
this is incorrect: e->user_data is only scanned during minor GC if it is the
right side of the 'young' pointer, and this might not be the case.

The fix is to use a normal CAMLlocal1 root for this value.

The testcase is tricky, but reliably crashes before and passes after this fix.