fix(multicore): Domain.join: do not lose backtrace
When an exception from a joined domain is raised, the backtrace is lost.
Instead, the backtrace the user gets indicates that it originated from
Domain.join, as opposed to where the error occurred inside the domain.
We (Semgrep) have started noticing this in CI runs, which makes pinning
the root cause difficult.

Note that this necessitated reordering `Domain` in stdlib/StdlibModules.