fix(windows): Fix racy socketpair on Windows
Previously socketpair on win32 relied on GetTempFileName to generate a
unix socket path. This API uses very little entropy and thus relies on
the created file essentially serving as a lock to avoid concurrent
calls picking the same temp name. However, since socketpair must
delete the file before binding to that path, it essentially
relinquishes the lock for that window. This sometimes leads to
address-in-use errors when concurrent threads or processes are trying
to create socketpairs. See ocaml/ocaml#14372.

This patch replaces that path name generation with one that uses the
process ID and an incrementing socketpair-call-specific ID. This rules
out the possibility of concurrent calls to socketpair picking the same
temp name. This does not rule out other threads or processes
deliberately colliding with the name, but the peerid check later
errors if another process collides.

Co-Authored-By: Nathan Taylor <nathan.taylor@semgrep.com>
(cherry picked from commit 5c8e0a24fe7d8320cd0e9174b9cd4f1cbbad154d)