ci: remove 'changes updated' workflow
We track changes of our fork, which are all applied to the 5.3.0-semgrep
branch, in the README on the trunk repo.  We keep it "out of tree"
because as we catch up with the upstream changes in e.g. the
in-development 5.5 compiler that we've backported here, we want to make
sure that reverting those commits don't revert the changelog.

What this does mean, though, is that the CHANGES file is somewhat
redundant, but CI can fail if it hasn't been updated.  So, this patch
simply removes that check from the workflow.  (It's still on us to
ensure that the README on trunk is still appropriately updated, though.)